Add PowerShell language
PowerShell is my day to day language and being able to automatically add the license to the header of every single file would be a god-send for me as it takes a lot of work. Please share the implementation with this pull request for all other users. Let's not restrict this to just the core compiled languages that we were all once used to.